% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gains.R
\name{add_tax_liability}
\alias{add_tax_liability}
\title{Add Tax Liability function}
\usage{
add_tax_liability(gain, st_taxrate = 0.3, lt_taxrate = 0.15)
}
\arguments{
\item{gain}{object of gains class}

\item{st_taxrate}{short term tax rate on gains for holings less than year}

\item{lt_taxrate}{long term tax rate on}
}
\value{
updated gains with tax impact
}
\description{
Appends the tax impact of a realized gain
}
