% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trades.R
\name{as.data.frame.trade}
\alias{as.data.frame.trade}
\title{Function to convert trade to data.frame}
\usage{
\method{as.data.frame}{trade}(x, ...)
}
\arguments{
\item{x}{trade object}

\item{...}{additional arguments. not currently implemented}
}
\description{
Function to convert trade to data.frame
}
