% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activity.R
\name{as_tibble.activity}
\alias{as_tibble.activity}
\title{Function to convert activity to tibble}
\usage{
as_tibble.activity(x)
}
\arguments{
\item{x}{trade object}

\item{...}{additional arguments. not currently implemented}
}
\description{
Function to convert activity to tibble
}
