% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gains.R
\name{as_tibble.gains}
\alias{as_tibble.gains}
\title{Function to convert gains to tibble}
\usage{
as_tibble.gains(x, ...)
}
\arguments{
\item{x}{trade object}

\item{...}{additional arguments. not currently implemented}
}
\description{
Function to convert gains to tibble
}
