% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Income.R
\name{as_tibble.income}
\alias{as_tibble.income}
\title{Function to convert income to tibble}
\usage{
as_tibble.income(x)
}
\arguments{
\item{x}{trade object}

\item{...}{additional arguments. not currently implemented}
}
\description{
Function to convert income to tibble
}
