% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trades.R
\name{as_tibble.trade}
\alias{as_tibble.trade}
\title{Function to convert trade to tibble}
\usage{
as_tibble.trade(x, ...)
}
\arguments{
\item{x}{trade object}

\item{...}{additional arguments. not currently implemented}
}
\description{
Function to convert trade to tibble
}
