% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates.R
\name{calc_returns}
\alias{calc_returns}
\title{Calculate Returns from Estiamates Object}
\usage{
calc_returns(eobj)
}
\arguments{
\item{eobj}{estimates object}
}
\description{
Calculate Returns from Estiamates Object
}
