% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{compare_constraints}
\alias{compare_constraints}
\title{Compare Constriant Checks for two portfolios}
\usage{
compare_constraints(pobj1, pobj2, cobj, eobj)
}
\description{
Function used to evaluate new portfolio to see if constraints improved upon
}
