% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activity.R
\name{deposit}
\alias{deposit}
\title{Create Deposit Helper function}
\usage{
deposit(date, amount, desc = "")
}
\arguments{
\item{date}{date of deposit. Date type}

\item{amount}{amount of deposit. numeric type}

\item{desc}{optional description of deposit. string type}
}
\value{
valid activity object
}
\description{
Function to create a deposit activity type
}
\examples{
deposit(Sys.Date(), 100, "monthly investment deposit")
}
