% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Income.R
\name{dividend}
\alias{dividend}
\title{Create Dividend Helper function}
\usage{
dividend(date, symbol, quantity, dividend, amount, desc = "")
}
\arguments{
\item{date}{date of dividend. Date type}

\item{symbol}{symbol ticker. character type}

\item{quantity}{number of shares. numeric type}

\item{dividend}{dividend per share. numeric type}

\item{amount}{dollar amount. numeric type}

\item{desc}{optional description of dividend. string type}
}
\value{
valid income object
}
\description{
Function to create a dividend income type
}
\examples{
dividend(Sys.Date(), symbol = "SPY", quantity = 10, dividend = 2, amount = 20)
}
