% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activity.R
\name{fee}
\alias{fee}
\title{Create Fee Activity Type Helper}
\usage{
fee(date, amount, desc = "")
}
\arguments{
\item{date}{date of activity. Date type}

\item{amount}{activity dollar amount. numeric type}

\item{desc}{activity description. string type}
}
\value{
valid activity object
}
\description{
Create a valid fee object of class Activity
}
\examples{
fee(Sys.Date(), 5, "transaction cost")
}
