% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{get_activity}
\alias{get_activity}
\title{Get Portfolio Activity}
\usage{
get_activity(pobj)
}
\arguments{
\item{pobj}{portfolio object}
}
\value{
activity object
}
\description{
getter function to return Portfolio's activity
}
\examples{
library(tidyverse)
portfolio("new_port", cash = 100) \%>\%
get_activity(.)
}
