% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{get_cash}
\alias{get_cash}
\title{Get Portfolio's Cash}
\usage{
get_cash(pobj)
}
\arguments{
\item{pobj}{portfolio object}
}
\value{
numeric cash value
}
\description{
getter function to return Portfolio's cash balance
}
\examples{
library(tidyverse)
portfolio("new_port") \%>\%
get_cash(.)
}
