% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Symbols.R
\name{get_current_prices}
\alias{get_current_prices}
\title{Get Current Security Prices}
\usage{
get_current_prices(symbols, error_handling = "pass", warning = FALSE)
}
\arguments{
\item{symbols}{vector of stock symbol characters. can be 1 or more}

\item{error_handling}{option to handle errors within foreach loop. options
are 'pass', 'remove', or 'stop'}

\item{warning}{logical argument to print getSymbol warnings to console}
}
\value{
data.frame with 1 record per symbol with current adjusted close price
}
\description{
Wrapper function around get_prices function that returns current prices. Adds
timestamp of date pulled
}
\examples{
\dontrun{
library(tidyverse)
library(quantmod)
symbols <- c("spy", "tlt")
prices <- symbols \%>\% get_current_prices(.)
}
}
