% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{get_gains}
\alias{get_gains}
\title{Get Portfolio Realized Gains}
\usage{
get_gains(pobj)
}
\arguments{
\item{pobj}{portfolio object}
}
\value{
Portfolio's realized gains
}
\description{
getter function to return Portfolio's realized gains
}
\details{
realized gains are gains or losses as a result of investment activity.
Realized when sold
}
\examples{
library(tidyverse)
portfolio("new_port") \%>\%
get_gains(.)
}
