% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{get_holding}
\alias{get_holding}
\title{Get Holding helper function}
\usage{
get_holding(pobj, .id)
}
\arguments{
\item{pobj}{portfolio object}

\item{.id}{id of holding to extract}
}
\value{
holding with id == .id
}
\description{
Get Holding helper function
}
