% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{get_holdings}
\alias{get_holdings}
\title{Get Portfolio Holdings}
\usage{
get_holdings(pobj)
}
\arguments{
\item{pobj}{portfolio object}
}
\value{
holdings tibble
}
\description{
getter function to return Portfolio's current holdings
}
\details{
holdings are currently held investment securities such as a stock or etf
}
\examples{
library(tidyverse)
portfolio("new_port", cash = 100) \%>\%
get_holdings(.)
}
