% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{get_holdings_market_value}
\alias{get_holdings_market_value}
\title{Get Portfolio's Holdings Market Value}
\usage{
get_holdings_market_value(pobj)
}
\arguments{
\item{pobj}{portfolio object}
}
\value{
tibble with portfolio holdings market value
}
\description{
Get Portfolio's Holdings Market Value
}
