% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{get_income}
\alias{get_income}
\title{Get Portfolio Investment Income}
\usage{
get_income(pobj)
}
\arguments{
\item{pobj}{portfolio object}
}
\value{
Portfolio's past income payments from investments
}
\description{
Returns realized income only. Does not estimate future income payments
}
\examples{
library(tidyverse)
portfolio("new_port") \%>\%
get_income(.)
}
