% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{get_market_value}
\alias{get_market_value}
\title{Get Portfolio Market Value}
\usage{
get_market_value(pobj)
}
\arguments{
\item{pobj}{portfolio object}
}
\value{
tibble with portfolio's market value
}
\description{
Get Portfolio Market Value
}
