% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio-past.R
\name{get_past_cash}
\alias{get_past_cash}
\title{Get Portfolio Past Cash Balance}
\usage{
get_past_cash(pobj)
}
\arguments{
\item{pobj}{portfolio object}
}
\value{
tibble with historical cash balance
}
\description{
Function to extract the full historical portfolio cash balance
}
