% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio-past.R
\name{get_past_dividends}
\alias{get_past_dividends}
\title{Get Portfolio Past Dividends}
\usage{
get_past_dividends(pobj)
}
\arguments{
\item{pobj}{portfolio object}
}
\value{
tibble with historical dividends. arrange by symbol and date with
  holding quantity, dividend and total payment amount
}
\description{
Function to extract the full historical portfolio dividends.
}
