% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio-past.R
\name{get_past_holdings}
\alias{get_past_holdings}
\title{Get Portfolio Past Holdings}
\usage{
get_past_holdings(pobj)
}
\arguments{
\item{pobj}{portfolio object}
}
\value{
tibble with historical holdings. arrange by symbol and date with
  holding quantity
}
\description{
Function to extract the full historical portfolio holdings. Returns expanded
date sequence of symbol holdings
}
