% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio-past.R
\name{get_past_holdings_market_value}
\alias{get_past_holdings_market_value}
\title{Get Portfolio Past Holdings Market Value}
\usage{
get_past_holdings_market_value(pobj)
}
\arguments{
\item{pobj}{portfolio object}
}
\value{
tibble with historical holdings. arrange by symbol and date with
  holding quantity, price and value
}
\description{
Function to extract the full historical portfolio holdings and calculate
market value. Returns expanded date sequence of symbol holdings with price and market value
}
