% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio-past.R
\name{get_past_tax_liability}
\alias{get_past_tax_liability}
\title{Get Potfolio Past Tax Liability}
\usage{
get_past_tax_liability(pobj)
}
\arguments{
\item{pobj}{portfolio object}
}
\value{
tibble with historical tax liability amount
}
\description{
Function to extact and calculate tax liability for past gains. Function
resets tax liability for each calander year
}
