% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{get_portfolio_returns}
\alias{get_portfolio_returns}
\title{Get Portfolio Return}
\usage{
get_portfolio_returns(pobj, start_date, end_date = Sys.Date())
}
\arguments{
\item{pobj}{portfolio object}

\item{start_date}{starting date for return horizon. Date class required}

\item{end_date}{end date for return horizon. Date class required}
}
\value{
data.frame with cash, investments, net investment and net value
  return statistic
}
\description{
Function extracts portfolio market values and calculates returns for given
time horizon. Function calculates the net investment value which is return on
invests only (net cash change)
}
