% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trades.R
\name{get_sell_ids}
\alias{get_sell_ids}
\title{Get Sell Trades}
\usage{
get_sell_ids(pobj, symbol, quantity)
}
\arguments{
\item{pobj}{portfolio object}

\item{symbols}{vector of holding symbols to filter sell tickets by}

\item{amount}{trade amount}

\item{partial}{logical option to allow for partial trade tickets with an
quantity less than the quantity argument. default is TRUE}
}
\value{
data.frame with possible sell trades
}
\description{
Returns all possible sell trades from a portfolio object
}
\details{
Checks current holdings and generates sell tickets based on symbol, trade
amount and lot_size
}
