% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{get_symbol_portfolio_share}
\alias{get_symbol_portfolio_share}
\title{Get Share of Total Portfolio By Symbol}
\usage{
get_symbol_portfolio_share(pobj)
}
\arguments{
\item{pobj}{portfolio object}
}
\value{
data.frame with portfolio share by symbol
}
\description{
Aggregates holdings portfolio share to symbol
}
