% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activity.R
\name{incur_fee}
\alias{incur_fee}
\title{Incur Fee function}
\usage{
incur_fee(pobj, date = Sys.Date(), amount, desc = "")
}
\arguments{
\item{pobj}{portfolio object}

\item{date}{date of activity. Date type}

\item{amount}{activity dollar amount. numeric type}

\item{desc}{activity description. string type}
}
\value{
updated portfolio object
}
\description{
incurr fee on a portfolio. fees could be transaction or management fees
}
\details{
subtracts fee amount from cash balance and adds record to actvity
}
\examples{
library(tidyverse)
portfolio("new_port", cash = 100) \%>\%
incur_fee(amount = 50)
}
