% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Income.R
\name{interest}
\alias{interest}
\title{Create Interest Helper function}
\usage{
interest(date, principal = 1, amount, desc = "")
}
\arguments{
\item{date}{date of interest payment. Date type}

\item{principal}{principal amount . numeric type}

\item{amount}{dollar amount of interest payment. numeric type}

\item{desc}{optional description of interest. string type}
}
\value{
valid income object
}
\description{
Function to create an interest income type
}
\examples{
interest(Sys.Date(), 1000, 1, "monthly interest")
}
