% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio-past.R
\name{intialize_portfolio}
\alias{intialize_portfolio}
\title{Initialize Portfolio from Activity List}
\usage{
intialize_portfolio(pobj, activity)
}
\arguments{
\item{pobj}{portfolio object}

\item{activity}{list of portfolio activity. Examples trade, income and other activity}
}
\value{
portfolio with activity added and past market value calculated
}
\description{
Initialize a portfolio from a list of prior activity. Function process all
activity and updates the historical portfolio market value
}
