% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{load_portfolio}
\alias{load_portfolio}
\title{Load Existing Portfolio}
\usage{
load_portfolio(path)
}
\arguments{
\item{path}{valid file path to porfolio object. should be .RData file type}
}
\value{
portfolio object
}
\description{
Function to load into memory saved portfolio object in .RData file
}
