% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palette.R
\name{madstork_cols}
\alias{madstork_cols}
\title{Function to extract madstork colors as hex codes}
\usage{
madstork_cols(...)
}
\arguments{
\item{...}{Character names of madstork_colors}
}
\description{
Function to extract madstork colors as hex codes
}
