% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palette.R
\name{madstork_pal}
\alias{madstork_pal}
\title{Return function to interpolate a madstork color palette}
\usage{
madstork_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in madstork_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a madstork color palette
}
