% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trades.R
\name{make_buy}
\alias{make_buy}
\title{Make Buy Trade Execution function}
\usage{
make_buy(
  pobj,
  date = Sys.Date(),
  symbol,
  quantity,
  price,
  desc = "",
  trans_cost = 0.05
)
}
\arguments{
\item{pobj}{portfolio object}

\item{date}{date of transaction. Date type}

\item{symbol}{symbol ticker. character type}

\item{quantity}{number of shares. numeric type}

\item{price}{price of shares. numeric type}

\item{desc}{trade description. string type}

\item{trans_cost}{transaction cost (dollars per share)}
}
\value{
updated portfolio object
}
\description{
Function executions a buy trade
}
\details{
subtracts trade amount from cash balance, updates portfolio internal trade
history and holdings
}
\examples{
library(tidyverse)
portfolio("new_port") \%>\%
make_deposit(Sys.Date(), amount = 2000) \%>\%
make_buy(symbol = "SPY", quantity = 10, price = 100)
}
