% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activity.R
\name{make_deposit}
\alias{make_deposit}
\title{Make Deposit function}
\usage{
make_deposit(pobj, deposit)
}
\arguments{
\item{pobj}{portfolio object}

\item{deposit}{deposit instance of activity object}
}
\value{
portfolio object with updated deposit
}
\description{
add deposit to portfolio
}
\details{
adds to cash balance and appends the deposit to the
portfolio activity
}
\examples{
library(tidyverse)
portfolio("new_port", cash = 100) \%>\%
make_deposit(deposit(date=Sys.Date(), amount = 100))
}
