% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trades.R
\name{make_sell}
\alias{make_sell}
\title{Make Sell Trade Execution function}
\usage{
make_sell(pobj, id, date = Sys.Date(), quantity, price, desc = "",
  trans_cost = 0.05)
}
\arguments{
\item{pobj}{portfolio object}

\item{id}{holding id to sell}

\item{date}{date of transaction. Date type}

\item{quantity}{number of shares. numeric type}

\item{price}{price of shares. numeric type}

\item{desc}{trade description. string type}

\item{trans_cost}{transaction cost (dollars per share)}
}
\value{
updated portfolio object
}
\description{
Function executions a sell trade
}
\details{
add trade amount to cash balance, updates portfolio internal trade
history and holdings
}
\examples{
library(tidyverse)
 p1 <- portfolio("new_port") \%>\%
       make_deposit(Sys.Date(), amount = 2000) \%>\%
       make_buy(Sys.Date()-1, symbol = "SPY", quantity = 10, price = 100) \%>\%
       make_sell(id = 1, quantity = 5, price = 105)
}
