% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activity.R
\name{make_withdraw}
\alias{make_withdraw}
\title{Make Withdraw function}
\usage{
make_withdraw(pobj, date = Sys.Date(), amount, desc = "")
}
\arguments{
\item{pobj}{portfolio object}

\item{date}{date of deposit. Date type}

\item{amount}{amount of deposit. numeric type}

\item{desc}{optional description of deposit. string type}
}
\value{
updated portfolio object
}
\description{
make withdraw from a portfolio
}
\details{
subtracts withdraw amount from cash balance and adds record to actvity
}
\examples{
library(tidyverse)
portfolio("new_port", cash = 100) \%>\%
make_withdraw(date=Sys.Date(), amount = 50)
}
