% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{nbto}
\alias{nbto}
\title{Next Best Trade Optimization}
\usage{
nbto(pobj, cobj, eobj, prices, trade_pairs, target, minimize, amount, lot_size)
}
\arguments{
\item{pobj}{portfolio object}

\item{cobj}{constraints object}

\item{eobj}{estimates object}

\item{prices}{data.frame with current symbol prices. Has to contain all
symbols included in estimates object. Should also include dividend}

\item{trade_pairs}{trade pairs to consider. trades limited to the trade pairs
provided}

\item{target}{target objective}

\item{minimize}{logical flag to minimize target objective}

\item{amount}{trade amount in dollars}

\item{lot_size}{lot size for trades}
}
\description{
Function to select optimal next best trade given the portfolio, estimates,
possible trades and constraints
}
\details{
Used in meet_constraint functions and optimize function
}
