% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activity.R
\name{new_activity}
\alias{new_activity}
\title{Activity Object Constructor function}
\usage{
new_activity(activity, date, amount, desc)
}
\arguments{
\item{activity}{activity type. string type.}

\item{date}{date of activity. Date type}

\item{amount}{activity dollar amount. numeric type}

\item{desc}{activity description. string type}
}
\value{
object of class activity
}
\description{
Function to create an object of class activity
}
