% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gains.R
\name{new_gains}
\alias{new_gains}
\title{Realized Gains Constructor function}
\usage{
new_gains(
  symbol,
  quantity,
  purchase_date,
  purchase_price,
  sale_date,
  sale_price,
  gain,
  type
)
}
\arguments{
\item{symbol}{symbol ticker. character type}

\item{quantity}{number of shares. numeric type}

\item{purchase_date}{date of purchase. Date type}

\item{purchase_price}{cost basis. numeric type}

\item{sale_date}{date of sale. Date type}

\item{sale_price}{sale price}

\item{gain}{gain amount. numeric type}

\item{type}{gain type. short or long term types. character type}
}
\value{
obj of class gains
}
\description{
Creates new object of class gains
}
