% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Income.R
\name{new_income}
\alias{new_income}
\title{Income Object Constructor function}
\usage{
new_income(type, date, symbol, quantity, payment, amount, desc)
}
\arguments{
\item{type}{trade type. string type.}

\item{date}{date of income payment. Date type}

\item{symbol}{symbol ticker. character type}

\item{quantity}{number of shares. numeric type}

\item{payment}{payment per share. Dividend for equity and interest for bond.
numeric type}

\item{amount}{income dollar amount. numeric type}

\item{desc}{income description. string type}
}
\value{
object of class income
}
\description{
Function to create an object of class income
}
