% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{new_portfolio}
\alias{new_portfolio}
\title{Portfolio Object Constructor function}
\usage{
new_portfolio(name, cash)
}
\arguments{
\item{name}{name of portfolio. requires string input}

\item{cash}{cash balance. requires numeric input}
}
\value{
portfolio object
}
\description{
Function creates a new S3 portfolio object
}
