% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trades.R
\name{new_trade}
\alias{new_trade}
\title{Trade Object Constructor function}
\usage{
new_trade(type, date, symbol, price, quantity, amount, desc)
}
\arguments{
\item{type}{trade type. string type.}

\item{date}{date of transaction. Date type}

\item{symbol}{symbol ticker. character type}

\item{price}{price of shares. numeric type}

\item{quantity}{number of shares. numeric type}

\item{amount}{trade dollar amount. numeric type}

\item{desc}{trade description. string type}
}
\value{
object of class activity
}
\description{
Function to create an object of trade class
}
