% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio-past.R
\name{past_dividend_income}
\alias{past_dividend_income}
\title{Past Dividend Income}
\usage{
past_dividend_income(pobj)
}
\arguments{
\item{pobj}{portfolio object}
}
\value{
updated portfolio object with all dividends processed and added to
  portfolio income
}
\description{
Function to process all past dividend income from historical holdings.
Functions removes any prior dividend income, gets all past dividends and then
updates portfolio
}
