% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio-past.R
\name{past_market_value}
\alias{past_market_value}
\title{Calculate Portfolio Past Market Value}
\usage{
past_market_value(pobj)
}
\arguments{
\item{pobj}{portfolio object}
}
\value{
updated portfolio object with all dividends processed and added to
  portfolio income
}
\description{
Function to calculate the complete portfolio historical market value.
Calculates the past investment value, cash balance, and tax liability
seperately
}
