% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{portfolio}
\alias{portfolio}
\title{Portfolio creation helper function}
\usage{
portfolio(name, cash = 0)
}
\arguments{
\item{name}{name of portfolio. requires string input}

\item{cash}{cash balance. requires numeric input. defaults to 0}
}
\value{
returns a Porftolio object
}
\description{
Function to create Portfolio object. Calls internal new_portfolio and
validate_portfolio functions
}
\examples{
portfolio("new_port", cash=0)
}
