% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activity.R, R/Income.R, R/Portfolio.R,
%   R/Trades.R
\name{process.activity}
\alias{process.activity}
\alias{process.income}
\alias{process}
\alias{process.trade}
\title{Process Portfolio Activity}
\usage{
\method{process}{activity}(obj, pobj, ...)

\method{process}{income}(obj, pobj, ...)

process(obj, pobj, ...)

\method{process}{trade}(obj, pobj, trans_cost = 0.05, ...)
}
\arguments{
\item{obj}{activity object to process}

\item{pobj}{portfolio object}

\item{...}{additional arguments to pass to the portfolio object}

\item{trans_cost}{transaction costs per share. default is 5cents ~ 0.05}
}
\value{
updated portfolio object with activity
}
\description{
Generic function to process madstork activity, income, and trade class
objects on a portfolio
}
