% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Income.R
\name{recieve_dividend}
\alias{recieve_dividend}
\title{Recieve Dividend payment}
\usage{
recieve_dividend(pobj, date = Sys.Date(), symbol, quantity = NULL,
  dividend = NULL, amount = NULL, desc = "")
}
\arguments{
\item{pobj}{portfolio object}

\item{date}{date of dividend. Date type}

\item{symbol}{symbol ticker. character type}

\item{quantity}{number of shares. numeric type}

\item{dividend}{dividend per share. numeric type}

\item{amount}{dollar amount. numeric type}

\item{desc}{optional description of dividend. string type}
}
\value{
updated portfolio object
}
\description{
Function updates cash balance and adds record to activity and income records
}
\examples{
library(tidyverse)
portfolio("new_port", cash = 2000) \%>\%
make_buy(symbol = "SPY", quantity = 10, price = 100) \%>\%
recieve_dividend(symbol = "SPY", quantity = 10, dividend = 2)
}
