% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Income.R
\name{recieve_interest}
\alias{recieve_interest}
\title{Recieve Interest payment}
\usage{
recieve_interest(pobj, date = Sys.Date(), amount, desc = "")
}
\arguments{
\item{pobj}{portfolio object}

\item{date}{date of interest payment. Date type}

\item{amount}{dollar amount of interest payment. numeric type}

\item{desc}{optional description of interest. string type}
}
\value{
updated portfolio object
}
\description{
Function updates cash balance and adds record to activity and income records
}
\examples{
library(tidyverse)
portfolio("new_port") \%>\%
make_deposit(amount = 2000) \%>\%
make_buy(symbol = "SPY", quantity = 10, price = 100) \%>\%
recieve_interest(amount = 5)
}
