% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reports.R
\name{report.portfolio}
\alias{report.portfolio}
\title{Portfolio Performance Report}
\usage{
\method{report}{portfolio}(portfolio, output_file, output_dir)
}
\arguments{
\item{portfolio}{portfolio object}

\item{output_file}{name of output file. Needs a html extension}

\item{output_dir}{directory to file folder to}
}
\description{
Function to create a Porftolio Performance Report for input Portfolio.
}
\details{
Uses portfolio_report.Rmd template in the madstork package
}
