% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reports.R
\name{report}
\alias{report}
\title{Generic Report Function}
\usage{
report(...)
}
\arguments{
\item{...}{additional parameters to pass to rmarkdown::render function}

\item{portfolio}{portfolio object}

\item{output_file}{name of output file. Needs a html extension}

\item{output_dir}{directory to write output file to}

\item{pandoc_dir}{pandoc directory path}
}
\description{
Generic Report Function
}
