% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reports.R
\name{risk_report}
\alias{risk_report}
\title{MadStork Risk Report Function}
\usage{
risk_report(portfolio, output_file = "risk-report.html",
  output_dir = getwd(),
  pandoc_dir = "C:/Program Files/RStudio/bin/pandoc", ...)
}
\arguments{
\item{portfolio}{portfolio object}

\item{output_file}{name of output file. Needs a html extension}

\item{output_dir}{directory to write output file to}

\item{pandoc_dir}{pandoc directory path}

\item{...}{additional parameters to pass to rmarkdown::render function}
}
\description{
Creates a Risk Report on Portfolio provided.
}
\details{
Uses risk_report.Rmd template in MadStork package
}
