% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{save_portfolio}
\alias{save_portfolio}
\title{Save Portfolio Function}
\usage{
save_portfolio(pobj, path, overwrite = TRUE)
}
\arguments{
\item{pobj}{portfolio object}

\item{path}{file path to save portfolio}

\item{overwrite}{logical object to overwrite existing portfolio if exists.
default to TRUE}
}
\description{
Function saves the portfolio to file. RData file extension required.
}
