% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palette.R
\name{scale_color_madstork}
\alias{scale_color_madstork}
\title{Color scale constructor for madstork colors}
\usage{
scale_color_madstork(palette = "main", discrete = TRUE,
  reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in madstork_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for madstork colors
}
