% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Income.R
\name{set_interest_rate}
\alias{set_interest_rate}
\title{Set Portfolio Interest Rate}
\usage{
set_interest_rate(pobj, rate)
}
\arguments{
\item{pobj}{portfolio object}

\item{rate}{interest rate. ex - `0.05 = 5%`}
}
\value{
updated portfolio object
}
\description{
Set the rate of interest portfolio earns on cash position
}
