% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{settle_tax_liability}
\alias{settle_tax_liability}
\title{Settle Portfolio Tax Liability}
\usage{
settle_tax_liability(pobj, date = Sys.Date(), amount, withdraw = FALSE)
}
\arguments{
\item{pobj}{portfolio object}

\item{date}{date of transaction. default is current date}

\item{amount}{amount of tax settlement}

\item{withdraw}{logical option to make a cash withdraw from portfolio}
}
\value{
updated portfolio object
}
\description{
Function to settle the tax liabilty. Option to make cash withdraw and add to
portfolio activity
}
\examples{
library(tidyverse)
portfolio("new_port", cash=0) \%>\%
 make_deposit(amount = 2000) \%>\%
 make_buy(Sys.Date()-1, symbol = "SPY", quantity = 10, price = 100) \%>\%
 make_sell(id = 1, quantity = 5, price = 105) \%>\%
 settle_tax_liability(amount = 7.5, withdraw = TRUE)
}
