% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{test_estimates}
\alias{test_estimates}
\title{Test Estimates}
\format{An estimates object.}
\usage{
test_estimates
}
\description{
Estimates object example used for testing. Symbols used are SPY, TLT, QQQ, and
GLD. The grain is set at 1 year with a 10 year lookback horizon. Sample mu
and Sigma calcuations added
}
\keyword{datasets}
