% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{test_prices}
\alias{test_prices}
\title{Test Prices}
\format{
An data.frame with with four variables: \code{last_updated}, \code{symbol},
  \code{price}, and \code{dividend}.
}
\usage{
test_prices
}
\description{
Current Prices provided for testings. Symbols used are SPY, TLT, QQQ, and
GLD.
}
\keyword{datasets}
