% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activity.R, R/Gains.R, R/Helpers.R, R/Income.R,
%   R/Trades.R
\name{to_tibble.activity}
\alias{to_tibble.activity}
\alias{to_tibble.gains}
\alias{to_tibble}
\alias{to_tibble.default}
\alias{to_tibble.income}
\alias{to_tibble.trade}
\title{Function to convert activity to tibble}
\usage{
\method{to_tibble}{activity}(x, ...)

\method{to_tibble}{gains}(x, ...)

to_tibble(x, ...)

\method{to_tibble}{default}(x, ...)

\method{to_tibble}{income}(x)

\method{to_tibble}{trade}(x, ...)
}
\arguments{
\item{x}{trade object}

\item{...}{additional arguments. not currently implemented}
}
\description{
Function to convert activity to tibble

Function to convert gains to tibble

Function to convert object to tibble

Default to tibble behavior

Function to convert income to tibble

Function to convert trade to tibble
}
