% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trades.R
\name{transfer}
\alias{transfer}
\title{Create Transfer Trade Helper}
\usage{
transfer(date, symbol, quantity, price, type = "", desc = "")
}
\arguments{
\item{date}{date of transaction. Date type}

\item{symbol}{symbol ticker. character type}

\item{quantity}{number of shares. numeric type}

\item{price}{price of shares. numeric type}

\item{type}{trade type. string type.}

\item{desc}{trade description. string type}
}
\value{
valid trade object
}
\description{
Create a valid transfer object of class Trade
}
\examples{
transfer(date = Sys.Date(), symbol = "SPY", quantity = 10, price = 100, type="out")
}
