% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trades.R
\name{transfer_in}
\alias{transfer_in}
\title{Transfer In Holdings function}
\usage{
transfer_in(pobj, date = Sys.Date(), symbol, quantity, price, desc = "")
}
\arguments{
\item{pobj}{portfolio object}

\item{date}{date of transaction. Date type}

\item{symbol}{symbol ticker. character type}

\item{quantity}{number of shares. numeric type}

\item{price}{price of shares. numeric type}

\item{desc}{trade description. string type}
}
\value{
updated portfolio object
}
\description{
Function adds holding from portfolio without a cash transaction
}
\examples{
 p1 <- portfolio("new_port", cash=0) \%>\%
       make_deposit(Sys.Date(), amount = 2000) \%>\%
       transfer_in(Sys.Date()-1, symbol = "SPY", quantity = 10, price = 100)
}
