% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio-update.R
\name{update_dividend_income}
\alias{update_dividend_income}
\title{Update Dividend Income}
\usage{
update_dividend_income(pobj)
}
\arguments{
\item{pobj}{portfolio object}
}
\value{
updated portfolio object
}
\description{
Auto dividend income function. Function searches for last dividend payment
for each symbol holding, gets all subsequent dividends and updates the
portfolio
}
\details{
Function to be used in daily portfolio-update script to remove need for
manual dividend updating
}
