% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{update_holdings_market_value}
\alias{update_holdings_market_value}
\title{Updated Porfolio Holding's Market Value}
\usage{
update_holdings_market_value(pobj, refresh = TRUE)
}
\arguments{
\item{pobj}{portfolio object}

\item{refresh}{logical option to refresh holdings price and dividend data.
default it TRUE}
}
\value{
data.frame with portfolio's holdings market value
}
\description{
Updateds current market price and annual dividends for portfolio holdings.
Calculates market value, income and yield
}
\examples{
\dontrun{
p1 <- portfolio("new_port", cash=0) \%>\%
 make_deposit(amount = 5000) \%>\%
 make_buy(Sys.Date()-365, symbol = "SPY", quantity = 10, price = 200) \%>\%
 make_buy(Sys.Date()-365, symbol = "TLT", quantity = 25, price = 100)
 update_holdings_market_value(p1)
}

}
