% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{update_market_value}
\alias{update_market_value}
\title{Update Porfolio Market Value}
\usage{
update_market_value(pobj, refresh = TRUE)
}
\arguments{
\item{pobj}{portfolio object}

\item{refresh}{logical option to refresh holdings price and dividend data.
default it TRUE}
}
\value{
Updated Porfolio object with new market value
}
\description{
Function to update porfolio and holding's market value
}
\details{
Function appends portfolio's market value record for historical analysis and
overwrites the holdings market value
}
\examples{
\dontrun{
p1 <- portfolio("new_port", cash=0) \%>\%
 make_deposit(amount = 5000) \%>\%
 make_buy(Sys.Date()-365, symbol = "SPY", quantity = 10, price = 200) \%>\%
 make_buy(Sys.Date()-365, symbol = "TLT", quantity = 25, price = 100)
 update_market_value(p1)
}
}
