% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activity.R
\name{validate_activity}
\alias{validate_activity}
\title{Activity Object Validation function}
\usage{
validate_activity(x)
}
\arguments{
\item{x}{object of class activity}
}
\value{
valid activity object
}
\description{
Function to validate activity input values
}
