% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gains.R
\name{validate_gains}
\alias{validate_gains}
\title{Gains Object Validation function}
\usage{
validate_gains(x)
}
\arguments{
\item{x}{object of class gains}
}
\value{
valid gains object
}
\description{
Function to validate gains input values
}
