% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Income.R
\name{validate_income}
\alias{validate_income}
\title{Income Object Validation function}
\usage{
validate_income(x)
}
\arguments{
\item{x}{object of class income}
}
\value{
valid income object
}
\description{
Function to validate income input values
}
