% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{validate_portfolio}
\alias{validate_portfolio}
\title{Portfolio Object Validation function}
\usage{
validate_portfolio(x)
}
\arguments{
\item{x}{portfolio object}
}
\value{
valid portfolio object
}
\description{
Validator function to check for valid inputs
}
