% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trades.R
\name{validate_trade}
\alias{validate_trade}
\title{Trade Object Validation function}
\usage{
validate_trade(x)
}
\arguments{
\item{x}{object of class trade}
}
\value{
valid trade object
}
\description{
Function to validate trade input values
}
