% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activity.R
\name{withdraw}
\alias{withdraw}
\title{Create Withdraw Helper function}
\usage{
withdraw(date, amount, desc = "")
}
\arguments{
\item{date}{date of deposit. Date type}

\item{amount}{amount of deposit. numeric type}

\item{desc}{optional description of deposit. string type}
}
\value{
valid activity object
}
\description{
Function to create a withdraw activity type
}
\examples{
withdraw(Sys.Date(), 100, "monthly withdraw")
}
