% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize function}
\usage{
capitalize(x, first_only = FALSE)
}
\arguments{
\item{x}{vector of character strings}

\item{first_only}{first word only}
}
\value{
vector of capitalized character strings
}
\description{
Function that capitalizes all words.
}
\examples{

}
