% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{get_activity}
\alias{get_activity}
\title{Get Portfolio Activity}
\usage{
get_activity(pobj)
}
\arguments{
\item{pobj}{portfolio object}
}
\value{
tibble of activity history
}
\description{
getter function to return Portfolio's activity
}
\details{
activity relates to any portfolio cash transactions
}
\examples{
library(tidyverse)
portfolio("new_port") \%>\%
make_deposit(amount = 100) \%>\%
get_activity(.)
}
