% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{get_holdings_market_value}
\alias{get_holdings_market_value}
\title{Get Porfolio Holding's Market Value}
\usage{
get_holdings_market_value(pobj)
}
\arguments{
\item{pobj}{}
}
\value{
data.frame with portfolio's holdings market value
}
\description{
Gets current market price and annual dividends for portfolio holdings.
Calculates market value, income and yield
}
\examples{
\dontrun{
p1 <- portfolio("new_port", cash=0) \%>\%
 make_deposit(amount = 5000) \%>\%
 make_buy(Sys.Date()-365, symbol = "SPY", quantity = 10, price = 200) \%>\%
 make_buy(Sys.Date()-365, symbol = "TLT", quantity = 25, price = 100)
get_holdings_market_value(p1)
}

}
