% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Symbols.R
\name{get_prices}
\alias{get_prices}
\title{Get Historical Security Prices}
\usage{
get_prices(
  symbols,
  start_date = "1990-01-01",
  end_date = Sys.Date(),
  error_handling = "pass",
  warning = FALSE
)
}
\arguments{
\item{symbols}{vector of stock symbol characters. can be 1 or more}

\item{start_date}{starting date for historical prices. default is
'1990-01-01'. Y-m-d format required}

\item{end_date}{ending date for historical prices. default is current date.
Y-m-d format required}

\item{error_handling}{option to handle errors within foreach loop. options
are 'pass', 'remove', or 'stop'}

\item{warning}{logical argument to print getSymbol warnings to console}
}
\value{
single tibble with symbol price history row bounded in long format
}
\description{
Wrapper function around get_ochlav function that returns historical securtity prices
}
\details{
Prices are set to adjusted close
}
\examples{
\donttest{
library(tidyverse)
library(quantmod)
symbols <- c("spy", "tlt")
prices <- symbols \%>\% get_prices(.)
}
}
