% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{get_tax_liability}
\alias{get_tax_liability}
\title{Get Portfolio Tax Liability}
\usage{
get_tax_liability(pobj)
}
\arguments{
\item{pobj}{portfolio object}
}
\value{
Porfolio's current tax liability
}
\description{
Get Portfolio Tax Liability
}
\examples{
library(tidyverse)
portfolio("new_port", cash = 100) \%>\%
get_tax_liability(.)
}
