% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trades.R
\name{make_sell}
\alias{make_sell}
\title{Make Sell Trade Execution function}
\usage{
make_sell(pobj, id, date, symbol, quantity, price, desc = "")
}
\arguments{
\item{pobj}{portfolio object}

\item{id}{trade id of holding to sell}

\item{date}{date of transaction. Date type}

\item{symbol}{symbol ticker. character type}

\item{quantity}{number of shares. numeric type}

\item{price}{price of shares. numeric type}

\item{desc}{trade description. string type}
}
\value{
updated portfolio object
}
\description{
Function executions a sell trade
}
\details{
add trade amount to cash balance, updates portfolio internal trade
history and holdings
}
\examples{
library(tidyverse)
 p1 <- portfolio("new_port", cash=0) \%>\%
       make_deposit(Sys.Date(), amount = 1000) \%>\%
       make_buy(Sys.Date()-1, symbol = "SPY", quantity = 10, price = 100) \%>\%
       make_sell(id = 1, date = Sys.Date(), symbol = "SPY", quantity = 5, price = 105)
}
