% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activity.R
\name{make_withdraw}
\alias{make_withdraw}
\title{Make Withdraw function}
\usage{
make_withdraw(pobj, withdraw)
}
\arguments{
\item{withdraw}{}
}
\description{
make withdraw from a portfolio
}
\details{
subtracts withdraw amount from cash balance and adds record to actvity
}
\examples{
portfolio("new_port", cash = 100) \%>\%
make_withdraw(withdraw(date=Sys.Date(), amount = 50)
}
