% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{portfolio}
\alias{portfolio}
\title{Portfolio Object Constructor function}
\usage{
portfolio(name, activity = NULL)
}
\arguments{
\item{name}{name of portfolio. requires string input}

\item{activity}{optional list of historical portfolio activity}
}
\value{
portfolio object
}
\description{
Function creates a new S3 portfolio object
}
