% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{save_portfolio}
\alias{save_portfolio}
\title{Save Portfolio Function}
\usage{
save_portfolio(pobj, dir = "./", overwrite = TRUE)
}
\arguments{
\item{pobj}{portfolio object}

\item{dir}{directory to save portfolio}

\item{overwrite}{logical object to overwrite existing portfolio if exists.
default to TRUE}
}
\description{
Function saves the portfolio as .RData file in specified directory. Function
uses portfolio name for file name
}
