% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trades.R
\name{transfer_out}
\alias{transfer_out}
\title{Transfer Out Holdings function}
\usage{
transfer_out(pobj, id, date = Sys.Date(), quantity, desc = "")
}
\arguments{
\item{pobj}{portfolio object}

\item{id}{holding id to sell}

\item{date}{date of transaction. Date type}

\item{quantity}{number of shares. numeric type}

\item{desc}{trade description. string type}
}
\value{
updated portfolio object
}
\description{
Function removes holding from portfolio without a cash transaction
}
\examples{
library(tidyverse)
 p1 <- portfolio("new_port") \%>\%
       make_deposit(Sys.Date(), amount = 2000) \%>\%
       make_buy(Sys.Date()-1, symbol = "SPY", quantity = 10, price = 100) \%>\%
       transfer_out(id = 1, quantity = 10)
}
